Move det_col and name_color_codes to constants
